Move session verification to Session
