+++ lib/phoenix_live_view/static.ex
@@ -2,7 +2,7 @@ defmodule Phoenix.LiveView.Static do
   # Holds the logic for static rendering.
   @moduledoc false
 
-  alias Phoenix.LiveView.{Socket, Utils, Diff, Session, Route}
+  alias Phoenix.LiveView.{Socket, Utils, Diff, Route}
 
   # Token version. Should be changed whenever new data is stored.
   @token_vsn 5
@@ -21,61 +21,6 @@ defmodule Phoenix.LiveView.Static do
   end
 
   def render(_other, _assigns), do: nil
-
-  @doc """
-  Verifies the session token.
-
-  Returns the decoded map of session data or an error.
-
-  ## Examples
-
-      iex> verify_session(AppWeb.Endpoint, "topic", encoded_token, static_token)
-      {:ok, %Session{} = decoded_session}
-
-      iex> verify_session(AppWeb.Endpoint, "topic", "bad token", "bac static")
-      {:error, :invalid}
-
-      iex> verify_session(AppWeb.Endpoint, "topic", "expired", "expired static")
-      {:error, :expired}
-  """
-  def verify_session(endpoint, topic, session_token, static_token) do
-    with {:ok, %{id: id} = session} <- verify_token(endpoint, session_token),
-         :ok <- verify_topic(topic, id),
-         {:ok, static} <- verify_static_token(endpoint, id, static_token) do
-      merged_session = Map.merge(session, static)
-      {live_session_name, vsn} = merged_session[:live_session] || {nil, nil}
-
-      session = %Session{
-        id: id,
-        view: merged_session.view,
-        root_view: merged_session.root_view,
-        parent_pid: merged_session.parent_pid,
-        root_pid: merged_session.root_pid,
-        session: merged_session.session,
-        assign_new: merged_session.assign_new,
-        live_session_name: live_session_name,
-        live_session_vsn: vsn,
-        # optional keys
-        router: merged_session[:router],
-        flash: merged_session[:flash]
-      }
-
-      {:ok, session}
-    end
-  end
-
-  defp verify_topic("lv:" <> session_id, session_id), do: :ok
-  defp verify_topic(_topic, _session_id), do: {:error, :invalid}
-
-  defp verify_static_token(_endpoint, _id, nil), do: {:ok, %{assign_new: []}}
-
-  defp verify_static_token(endpoint, id, token) do
-    case verify_token(endpoint, token) do
-      {:ok, %{id: ^id}} = ok -> ok
-      {:ok, _} -> {:error, :invalid}
-      {:error, _} = error -> error
-    end
-  end
 
   @doc """
   Verifies a LiveView token.
